+++ hsuanwu/env/minigrid/__init__.py
@@ -45,14 +45,14 @@ class TorchVecEnvWrapper(gym.Wrapper):
 
     def reset(self, **kwargs) -> Tuple[Tensor, Dict]:
         obs, info = self.env.reset(**kwargs)
-        obs = torch.as_tensor(obs, dtype=torch.float32, device=self._device)
+        obs = torch.as_tensor(obs, device=self._device)
         return obs, info
 
     def step(self, action: Tensor) -> Tuple[Tensor, Tensor, Tensor, bool, Dict]:
         obs, reward, terminated, truncated, info = self.env.step(
             action.squeeze(1).cpu().numpy()
         )
-        obs = torch.as_tensor(obs, dtype=torch.float32, device=self._device)
+        obs = torch.as_tensor(obs, device=self._device)
         reward = torch.as_tensor(
             reward, dtype=torch.float32, device=self._device
         ).unsqueeze(dim=1)
